feature/86cw10vv5: Moved DB Helper functions and added docs to function
