+++ behind-the-veil-siteroot/imports/ui/components/util.jsx
@@ -5,12 +5,8 @@
  */
 
 import {useState} from "react";
-import {useSubscribe, useTracker} from "meteor/react-meteor-data";
 import {Tracker} from "meteor/tracker";
 import {Meteor} from "meteor/meteor";
-import ServiceCollection from "../../api/collections/services";
-import ImageCollection from "../../api/collections/images";
-import UserCollection from "../../api/collections/users";
 
 /**
  * Retrieves current logged-in user's information
@@ -67,77 +63,3 @@ export function getUserInfo() {
     console.log("Current logged in user:" + JSON.stringify(userInfo))
     return userInfo;
 }
-
-/**
- * Used for a list of services (not for one specific service, or else its not as optimised (?)
- *
- * @param service_publication
- * @param params
- * @param filter
- * @returns {*[]}
- */
-export function getServices(service_publication, params, filter, requireArtist = false) {
-
-    console.log("service_publication", service_publication)
-    console.log("params", params)
-    console.log("filter", filter)
-
-    // get service data from database
-    const isLoadingUserServices = useSubscribe(service_publication, ...params);
-    let servicesData = useTracker(() => {
-        return ServiceCollection.find(filter).fetch();
-    });
-
-    const isLoadingServiceImages = useSubscribe('service_images', []);
-    let imagesData = useTracker(() => {
-        return ImageCollection.find({imageType: "service"}).fetch();
-    });
-
-    let isLoadingArtists = () => false
-    let artistsData = [];
-
-    if (requireArtist) {
-        isLoadingArtists = useSubscribe('all_artists');
-        artistsData = useTracker(() => {
-            return UserCollection.find({"profile.type": "artist"}).fetch();
-        });
-    }
-
-    const isLoading = isLoadingUserServices() || isLoadingArtists() || isLoadingServiceImages();
-
-    // manual aggregation of each service with their image
-    for (let i = 0; i < servicesData.length; i++) {
-        let foundImageMatch = false;
-
-        // aggregate with artist first
-        for (let j = 0; j < artistsData.length; j++) {
-
-            // find matching artist and add their name
-            if (servicesData[i].artistUsername === artistsData[j].username) {
-                servicesData[i].artistAlias = artistsData[j].profile.alias;
-                break;
-            }
-        }
-
-        // then aggregate with the FIRST image that belong to it
-        for (let j = 0; j < imagesData.length; j++) {
-
-            // find matching image for the service
-            if (imagesData[j].imageType === "service" && servicesData[i]._id === imagesData[j].target_id) {
-                servicesData[i].serviceImageData = imagesData[j].imageData;
-                foundImageMatch = true;
-                break;
-            }
-        }
-
-        // if not found any images, replace with default
-        if (!foundImageMatch) {
-            servicesData[i].serviceImageData = "/imageNotFound.png";
-        }
-    }
-    console.log("servicesData", servicesData)
-
-    return [isLoading, servicesData]
-}
-
-
